+++ CHANGELOG.md
@@ -10,7 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added examples.
 
 ### Changed
-- Schema format `callback` has been renamed to `process-graph`.
+- The JSON Schema keyword `format` has been replaced with the custom keyword `subtype`.
+- Schema format/subtype `callback` has been renamed to `process-graph`.
 - `clip`: Works on a single value instead on arrays (replaced parameter `data` with `x`). (#75)[https://github.com/Open-EO/openeo-processes/issues/75]
 - `debug`: Replaced with a completely new definition. (#82)[https://github.com/Open-EO/openeo-processes/issues/71], (#100)[https://github.com/Open-EO/openeo-api/issues/100], (#214)[https://github.com/Open-EO/openeo-api/issues/214]